requirements.txt
anthropic>=0.97
python-dotenv>=1.0
playwright>=1.49
